% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSGS_mainFun.R
\name{BSGS}
\alias{BSGS}
\title{Bayesian Gene Scoring System (BSGS)}
\usage{
BSGS(
  data,
  label_col,
  iterations = 1000,
  min_genes = round(floor(sqrt(length(colnames(data)[-1])))/2),
  max_genes = 5 * round(floor(sqrt(length(colnames(data)[-1])))),
  alpha = 0.3,
  n_genes = 150,
  plot_distribution = FALSE,
  save_plot = FALSE,
  file_path = "Results from the Bayesian Gene Scoring System (BSGS).png"
)
}
\arguments{
\item{data}{A data frame containing gene expression data.}

\item{label_col}{A string specifying the column name for the class labels in `data`.}

\item{iterations}{An integer specifying the number of iterations for the Bayesian updates (default = 10).}

\item{min_genes}{The minimum number of genes to sample in each iteration
(default = `round(floor(sqrt(number of genes)) / 2)`).}

\item{max_genes}{The maximum number of genes to sample in each iteration
(default = `5 * round(floor(sqrt(number of genes)))`).}

\item{alpha}{Weighting factor for the Bayesian update, balancing per-gene scores and subset scores (default = 0.3).}

\item{n_genes}{Number of top-ranked genes to return if no threshold is provided (default = 150).}

\item{plot_distribution}{Logical; if `TRUE`, generates a plot of the gene score distribution (default = FALSE).}

\item{save_plot}{Logical; if `TRUE`, saves the gene score distribution plot to a file (default = FALSE).}

\item{file_path}{A string specifying the file path for saving the gene score distribution plot
(default = "Results from the Bayesian Gene Scoring System (BSGS).png").}
}
\value{
A list with the following components:
\describe{
  \item{`final_genes`}{A character vector of the selected genes.}
  \item{`final_scores`}{A numeric vector of the scores corresponding to the selected genes.}
  \item{`posteriors`}{A numeric vector of the final posterior probabilities for all genes.}
  \item{`gene_scores`}{A numeric vector of the final gene scores for all genes.}
}
}
\description{
Implements a Bayesian-based feature selection algorithm for identifying significant genes.
This function performs iterative Bayesian updates to calculate gene scores and select
the most informative genes for classification tasks.
}
\examples{
\dontrun{
# Example usage
full_data <- as.data.frame(LUAD_MYE_MRS_GSE97168)
full_data$target <- as.factor(LUAD_MYE_MRS_target)
bsgs_results <- BSGS(
  data = full_data,
  label_col = "target",
  iterations = 1000,
  alpha = 0.3,
  min_genes = round(floor(sqrt(length(colnames(full_data[,-1])))) / 2),
  max_genes = 5 * round(floor(sqrt(length(colnames(full_data[,-1])))), 2),
  n_genes = 150,
  plot_distribution = TRUE,
  save_plot = TRUE,
  file_path = "./bsgs_results.png"
)
print(bsgs_results$final_genes)
}
}
