\name{Crisis2008}
\alias{Crisis2008}
\title{A cross-sectional data set from Rose and Spiegel.}
\description{A cross-sectional data set from Rose and Spiegel (2011), which is available at
\url{http://faculty.haas.berkeley.edu/arose}.  The response varialbe is 2008-2009 growth rate 
for the cirsis measure. This dataset consisits of 119 explanatory factors for the 
crisis for as many as 107 countries, but there are data missing for a number of countries. 
}
\usage{data(Crisis2008)}


