\name{Crisis2008BalancedData}
\alias{Crisis2008BalancedData}
\title{A cross-sectional data set from Rose and Spiegel with the removal of missing values.}
\description{A cross-sectional data set from Rose and Spiegel (2011), which is available at
\url{http://faculty.haas.berkeley.edu/arose}.  The response varialbe is 2008-2009 growth rate
for the cirsis measure. Rose and Spiegel originally consider 119 explanatory factors for the
crisis for as many as 107 countries, but there are data missing for a number of countries.
To maintain a balanced data set, we use 51 regressors for a sample of 72 countries.
These regressors can be classified into the nine theoretical groups of the crisis' origin
(the number in parentheses indicates the number of variables considered in the group):
principal factors (10), financial policies (three), financial conditions (four),
asset price appreciation (two), macroeconomic policies (four), institutions (11),
geography (four), financial linkages (one), and trade linkages (12). }

\usage{data(Crisis2008BalancedData)}



