% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LUAD_MYE_MRS_GSE97168.R
\docType{data}
\name{LUAD_MYE_MRS_GSE97168}
\alias{LUAD_MYE_MRS_GSE97168}
\title{LUAD_MYE_MRS_GSE97168 Dataset}
\format{
A list with two elements:
\describe{
  \item{LUAD_MYE_MRS_GSE97168}{A data frame with gene expression data.}
  \item{LUAD_MYE_MRS_target}{A factor vector with labels corresponding to tissue types.}
}
}
\usage{
data(LUAD_MYE_MRS_GSE97168)
}
\description{
Gene expression dataset and labels for LUAD_MYE_MRS_GSE97168.
}
\examples{
data(LUAD_MYE_MRS_GSE97168)
head(LUAD_MYE_MRS_GSE97168$LUAD_MYE_MRS_GSE97168)
head(LUAD_MYE_MRS_GSE97168$LUAD_MYE_MRS_target)
}
\keyword{datasets}
