\name{MSE.BSGS}
\alias{MSE.BSGS}
%\keyword{arith}
\title{Mean square error (MSE).}
\description{Calculate the mean square error for the sparse group selection problems.}
\usage{MSE.BSGS(Output, Y, X)}
\arguments{
\item{Output}{A list of random samples generated by either ``BSGS.Simple'' or ``BSGS.Sample'' for parameters from spare group selection problems.}
\item{Y}{Observations.}
\item{X}{Design matrix.}
}
\value{Reture the mean square error.}
\examples{
\dontrun{
output = BSGS.Simple(Y, X, Group.Index, r.value, eta.value, beta.value, tau2.value, 
	     rho.value, theta.value, sigma2.value, nu, lambda, Num.of.Iter.Inside.CompWise, 
	     Num.Of.Iteration, MCSE.Sigma2.Given)
MSE.BSGS(output, Y, X)
}

}

