\name{MSE.CGS.SMP}
\alias{MSE.CGS.SMP}
%\keyword{arith}
\title{Mean square error (MSE).}
\description{Calculate the mean square error.}
\usage{MSE.CGS.SMP(Output, Y, X)}
\arguments{
\item{Output}{A list of random samples for parameters generated by CGS or SMP algorithm.}
\item{Y}{Observations.}
\item{X}{Design matrix.}
}
\value{Reture the mean square error.}
\examples{
\dontrun{
output = BSGS.Simple(Y, X, Group.Index, r.value, eta.value, beta.value, tau2.value, 
		 rho.value, theta.value, sigma2.value, nu, lambda, Num.of.Iter.Inside.CompWise, 
		 Num.Of.Iteration, MCSE.Sigma2.Given)
MSE.CGS.SMP(output, Y, X)
}

}

