% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{normalize}
\alias{normalize}
\title{Normalize Gene Expression Data}
\usage{
normalize(x, gamma = 3)
}
\arguments{
\item{x}{Numeric vector of gene expression data.}

\item{gamma}{Power scaling factor (default = 3).}
}
\value{
A numeric vector with normalized values.
}
\description{
Applies min-max normalization and power scaling to gene expression data.
}
