% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{plot_final_scores_distribution}
\alias{plot_final_scores_distribution}
\title{Plot Final Scores Distribution}
\usage{
plot_final_scores_distribution(gene_scores, save_plot = FALSE, file_path)
}
\arguments{
\item{gene_scores}{A numeric vector of gene scores.}

\item{save_plot}{Logical. Whether to save the plot (default: FALSE).}

\item{file_path}{Character. File path for saving the plot (default: NULL).}
}
\value{
A ggplot object of the distribution of gene scores.
}
\description{
Generates a histogram of gene scores and overlays key statistics.
}
