% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{sample_genes}
\alias{sample_genes}
\title{Sample Genes Based on Posterior Probabilities}
\usage{
sample_genes(posteriors, num_samples)
}
\arguments{
\item{posteriors}{Numeric vector of posterior probabilities.}

\item{num_samples}{Number of samples to select.}
}
\value{
Indices of selected genes.
}
\description{
Selects a subset of genes based on posterior probabilities.
}
