% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{update_posteriors}
\alias{update_posteriors}
\title{Update Posterior Probabilities}
\usage{
update_posteriors(priors, likelihoods, floor_val = 1e-06)
}
\arguments{
\item{priors}{Numeric vector of prior probabilities.}

\item{likelihoods}{Numeric vector of likelihood values.}

\item{floor_val}{Minimum allowable posterior value (default = 1e-6).}
}
\value{
Updated posterior probabilities.
}
\description{
Updates posterior probabilities using priors and likelihoods.
}
